# src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: "1.0.0"
  title: "Sample API"
  description: Initial example to submit Elasticsearch queries

paths:
  /search:
    get:
      summary: "Submits a query to Elasticsearch"
      operationId: search
      tags:
        - search
      parameters:
        - name: query
          in: query
          description: Query to be submitted
          schema:
            type: string
        - name: page
          in: query
          description: Page number of results
          schema:
            type: integer
            default: 1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /search/image:
    post:
      summary: "Submits an image to identify a formula and search"
      operationId: searchByImage
      tags:
        - search
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        500:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    SearchResponse:
      type: object
      properties:
        totalPages:
          type: integer
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/Result"
    Result:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        abs:
          type: string
        formulas:
          type: array
          items:
            type: string
        formulas_latex:
          type: array
          items:
            type: string
    Error:
      type: object
      required:
        - message
      properties:
        message:
          description: A human readable error message
          type: string